add rigid body class
